Refactored code.
Code that should be shared between libcvmfs and the FUSE module
has been moved into cvmfs_common.cc.  This code is mostly
copied verbatim from the FUSE-independent code in cvmfs.cc.
Some small changes have been made to make it useful for libcvmfs.
The duplicated code has _not_ been removed from cvmfs.cc; doing
that is postponed for the future in order to minimize the
disturbance to the CVMFS trunk.